clutter/text: Check if attributes are equal before applying
Pango doesn't make it easy but the lists are generally extremely
short and it's worth the effort. Because when they are equal we
can avoid the `clutter_actor_queue_relayout` in
`clutter_text_set_attributes`. This particularly avoids stuttering
when moving the mouse over the gnome-shell calendar (which repeatedly
sets `font-feature-settings: "tnum"` on `calendar-day-base` themed
widgets).

Closes: https://gitlab.gnome.org/GNOME/gnome-shell/issues/1411

https://gitlab.gnome.org/GNOME/mutter/merge_requests/983